requirements.txt
aiogram==3.22.0
aiosqlite==0.21.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
alembic==1.13.2
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.3.4
typing_extensions==4.12.2
greenlet==3.0.3
APScheduler==3.10.4
httpx==0.27.2
magic_filter==1.0.12
aiofiles==23.2.1
asyncpg==0.29.0
pillow==11.3.0
